+++ app.js
@@ -7,9 +7,7 @@ const bodyParser = require('body-parser');
 const cors = require('cors');
 const path = require('path');
 const imagesPath = path.join(__dirname, 'images');
-const ftp = require('ftp');
-const fs = require('fs');
-const AdmZip = require('adm-zip');
+const cronScriptFtp = require('./script/cronScriptFtp');
 
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
@@ -90,6 +88,8 @@ app.use(errorController.logErrors);
 app.use(errorController.get404);
 
 
+cronScriptFtp.ftpConnect();
+
 // MONGODB -Connexion à la base de données
 mongoose.connect(`mongodb+srv://${process.env.DB_USER}:${process.env.DB_PASSWORD}@${process.env.DB_CLUSTER}/${process.env.DB_NAME}?retryWrites=true&w=majority`)
   .then(() => {
@@ -102,101 +102,4 @@ mongoose.connect(`mongodb+srv://${process.env.DB_USER}:${process.env.DB_PASSWORD
   .catch(err => {
     console.log('La connexion à la base de données a échoué', err);
   });
-
-// FTP CONNEXION (dependencies: node-ftp) - Connexion à ftp.solusoft-erp.com (+ Création des sous-dossiers Solusoft + Décompression des fichiers)
-const client = new ftp();
-const config = {
-  host: 'ftp.solusoft-erp.com',
-  port: 21,
-  user: 'Ricart@solusoft-erp.com',
-  password: 'ric2024art'
-};
-
-client.connect(config);
-
-// FTP Active l'écoute des événements et déclanche les callbacks.
-client.on('ready', () => {
-  client.list((err, list) => {
-    if (err) throw err;
-    // Ajout du fichier temporaire pour recevoir le WriteStream
-
-    const fs = require('fs');
-    const path = require('path');
-    
-    const folderPath = './folder';
-    
-    // Vérification SI dossier "solusoft/compressedFiles" existe déjà, sinon le créé.
-    fs.access("solusoft/compressedFiles", fs.constants.F_OK, (err) => {
-        if (err) {
-            // Directory does not exist, so create it
-            fs.mkdir("solusoft/compressedFiles", { recursive: true }, (err) => {
-                if (err) {
-                    console.error('Erreur de création du dossier "solusoft/compressedFiles":', err);
-                } else {
-                    console.log('Dossier "solusoft/compressedFiles" créé avec succès"');
-                }
-            });
-        } else {
-            console.log('Le dossier "solusoft/compressedFiles" existe déjà');
-        }
-    });
-
-        // Vérification SI dossier "solusoft/uncompressedFiles" existe déjà, sinon le créé.
-        fs.access("solusoft/uncompressedFiles", fs.constants.F_OK, (err) => {
-          if (err) {
-              // Directory does not exist, so create it
-              fs.mkdir("solusoft/uncompressedFiles", { recursive: true }, (err) => {
-                  if (err) {
-                      console.error('Erreur de création du dossier "solusoft/uncompressedFiles":', err);
-                  } else {
-                      console.log('Dossier "solusoft/uncompressedFiles" créé avec succès"');
-                  }
-              });
-          } else {
-              console.log('Le dossier "solusoft/uncompressedFiles" existe déjà');
-          }
-      });
-    
-
-    const pathName = 'solusoft/compressedFiles/downloadedFile.zip';
-
-    // FTP Sélection du fichier à télécharger sur le serveuer - produitTest.zip (temporaire)
-    const remoteFilePath = '/Produits/produitTest.zip'; // Fichier sur le FTP
-    //GET - Méthode pour télécharger un fichier
-    client.get(remoteFilePath, (err, stream) => {
-      if (err) {
-        console.error('Error downloading file:', err);
-        return;
-      }
-
-      // FTP PIPE PROCESS - WriteStream du fichier Serveur vers Stream Local
-      stream.pipe(fs.createWriteStream(pathName));
-
-
-      // FTP CLOSE EVENT - Écoute de la terminaison du processus de téléchargement
-      stream.once('close', () => {
-        console.log('fichier "downloadedFile.zip" téléchargé avec succès');
-        client.end(); // Close the FTP connection
-
-      // ADM-ZIP - Gestionnaire de compression/décompression de fichier
-      // ZIP - Chemin d'accès vers le fichier zip (produits)
-      const zipFilePath = 'solusoft/compressedFiles/downloadedFile.zip';
-
-
-      // Crée une instance d'AdmZip
-      const zip = new AdmZip(zipFilePath);
-
-
-      // Extrait le contenu du fichier compressé (produits)
-      zip.extractAllTo('solusoft/uncompressedFiles', true);
-
-      });
-    });
-  });
-});
-
-// FTP Gestion des événements
-client.on('error', (err) => {
-  console.log('FTP error occurred: ' + err);
-});
