+++ package.json
@@ -10,6 +10,7 @@
     "jsonwebtoken": "9.0.2",
     "mongodb": "6.2.0",
     "mongoose": "7.6.8",
+    "node-cron": "3.0.3",
     "nodemailer": "6.9.10",
     "nodemon": "3.0.1"
   },
